Update JS file
Seemed to be causing errors in Udacity testing
