--- JS/destinations.js
@@ -1,95 +1,0 @@
-const destinations = {   
-    japan: {
-        name: `Japan`,
-        info: `Japan is an island country in East Asia, located in the northwest Pacific Ocean. It is bordered on the west by the Sea of Japan, and extends from the Sea of Okhotsk in the north toward the East China Sea and Taiwan in the south.`,
-        fact1: `The world’s oldest company is in Japan`,
-        fact2: `It has the 11th largest population in the world`,
-        fact3: `There is 1 vending machine for every 24 people`,
-        image: `./Assets/japan.jpg`
-    },
-    southafrica: {
-        name: `South Africa`,
-        info: `South Africa is a country on the southernmost tip of the African continent, marked by several distinct ecosystems. Inland safari destination Kruger National Park is populated by big game. The Western Cape offers beaches, lush winelands around Stellenbosch and Paarl, craggy cliffs at the Cape of Good Hope, forest and lagoons along the Garden Route, and the city of Cape Town, beneath flat-topped Table Mountain.`,
-        fact1: `The longest continuous wine route on earth is found in South Africa`,
-        fact2: `It is the largest meat producer in Africa`,
-        fact3: `Table Mountain is one of the oldest mountains on the planet`,
-        image: `./Assets/south_africa.jpg`
-    },
-    scotland: {
-        name: `Scotland`,
-        info: `Scotland is a country that is part of the United Kingdom. Covering the northern third of the island of Great Britain, mainland Scotland has a 96-mile (154 km) border with England to the southeast and is otherwise surrounded by the Atlantic Ocean to the north and west, the North Sea to the northeast and the Irish Sea to the south`,
-        fact1: `The official animal of Scotland is the Unicorn`,
-        fact2: `Scotland has approximately 790 islands`,
-        fact3: `There are over 600 square miles of freshwater lakes`,
-        image: `./Assets/scotland.jpg`
-    },
-    iceland: {
-        name: `Iceland`,
-        info: `Iceland, a Nordic island nation, is defined by its dramatic landscape with volcanoes, geysers, hot springs and lava fields. Massive glaciers are protected in Vatnajökull and Snæfellsjökull national parks.`,
-        fact1: `Iceland was the last place on earth to be settled by humans`,
-        fact2: `About 11% of Iceland is covered by glaciers`,
-        fact3: `You can swim outdoors in hot springs all year round`,
-        image: `./Assets/iceland.jpg`
-    },
-    newzealand: {
-        name: `New Zealand`,
-        info: `New Zealand is an island country in the southwestern Pacific Ocean. It consists of two main landmasses—the North Island and the South Island —and more than 700 smaller islands, covering a total area of 268,021 square kilometres.`,
-        fact1: `It has the town with the longest name in the world`,
-        fact2: `There are more sheep than people`,
-        fact3: `It has the most southerly capital in the world`,
-        image: `./Assets/new_zealand.jpg`
-    },
-    australia: {
-        name: `Australia`,
-        info: `Australia, officially the Commonwealth of Australia, is a sovereign country comprising the mainland of the Australian continent, the island of Tasmania, and numerous smaller islands. It is the largest country by area in Oceania and the world's sixth-largest country.`,
-        fact1: `The Australian Alps get more snow than the Swiss Alps`,
-        fact2: `90% of Australians live on the coast`,
-        fact3: `The Great Barrier Reef is the largest eco-system in the world`,
-        image: `./Assets/australia.jpg`
-    },
-    thailand: {
-        name: `Thailand`,
-        info: `Thailand is a Southeast Asian country. It's known for tropical beaches, opulent royal palaces, ancient ruins and ornate temples displaying figures of Buddha. In Bangkok, the capital, an ultramodern cityscape rises next to quiet canalside communities and the iconic temples of Wat Arun, Wat Pho and the Emerald Buddha Temple (Wat Phra Kaew).`,
-        fact1: `Siamese cats are native to Thailand`,
-        fact2: `It is illegal to leave your house without underwear on`,
-        fact3: `Thailand is the world's largest exporter of rice`,
-        image: `./Assets/thailand.jpg`
-    },
-    france: {
-        name: `France`,
-        info: `France, in Western Europe, encompasses medieval cities, alpine villages and Mediterranean beaches. Paris, its capital, is famed for its fashion houses, classical art museums including the Louvre and monuments like the Eiffel Tower.`,
-        fact1: `France Is The Most-Visited Country In The World`,
-        fact2: `France Is Smaller Than Texas`,
-        fact3: `The French Eat 25,000 Tons Of Snails Each Year`,
-        image: `./Assets/france.jpg`
-    },
-    italy: {
-        name: `Italy`,
-        info: `Italy, a European country with a long Mediterranean coastline, has left a powerful mark on Western culture and cuisine. Its capital, Rome, is home to the Vatican as well as landmark art and ancient ruins.`,
-        fact1: `Italy has a free wine fountain`,
-        fact2: `All three of Europe’s active volcanoes are in Italy`,
-        fact3: `Italians invented pizza in Naples`,
-        image: `./Assets/italy.jpg`
-    },
-    spain: {
-        name: `Spain`,
-        info: `Spain, a country on Europe’s Iberian Peninsula, includes 17 autonomous regions with diverse geography and cultures. Capital city Madrid is home to the Royal Palace and Prado museum, housing works by European masters.`,
-        fact1: `Spain discovered oranges and chocolate`,
-        fact2: `Spain gave the first modern novel to the world`,
-        fact3: `The Spanish anthem has no words`,
-        image: `./Assets/spain.jpg`
-    },
-    portugal: {
-        name: `Portugal`,
-        info: `Portugal is a southern European country on the Iberian Peninsula, bordering Spain. Its location on the Atlantic Ocean has influenced many aspects of its culture: salt cod and grilled sardines are national dishes, the Algarve's beaches are a major destination and much of the nation’s architecture dates to the 1500s–1800s, when Portugal had a powerful maritime empire`,
-        fact1: `Portuguese Is the Official Language of Nine Countries`,
-        fact2: `Portugal Is the Oldest Country in Europe`,
-        fact3: `The Oldest Bookstore in the World Is in Portugal's Capital`,
-        image: `./Assets/portugal.jpg`
-    },
-    none: {
-        name: `--Please select an option--`,        
-    }
-}
-
-export { destinations }
