--- JS/userdata.js
@@ -1,11 +1,0 @@
-const userData = {
-    name : ``,
-    place1 : ``,
-    place2 : ``,
-    place3 : ``,
-    place4 : ``,
-    place5 : ``,
-    color : ``
-};
-
-export { userData }
